+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to the library should be put here
 
 ## Unreleased ~ 0.1.2 | 0.2.0
+ - Implement Display for JanetType
 
 ## 0.1.0 ~ 0.1.1
  - Basic Janet types manipulation